Conflicts:
	sprint2/target/classes/.gitignore
	sprint2/target/classes/common/database/DatabaseWrapper.class
	sprint2/target/classes/common/modelo/Pedido.class
	sprint2/target/classes/common/modelo/Producto.class
	sprint2/target/classes/usuario/controllers/BusinessLogicUtil.class
	sprint2/target/classes/usuario/views/VentanaDireccion$1.class
	sprint2/target/classes/usuario/views/VentanaDireccion$3.class
	sprint2/target/classes/usuario/views/VentanaDireccion.class
	sprint2/target/classes/usuario/views/VentanaInicio$1.class
	sprint2/target/classes/usuario/views/VentanaInicio$2.class
	sprint2/target/classes/usuario/views/VentanaInicio.class
	sprint2/target/classes/usuario/views/VentanaPago$3.class
	sprint2/target/classes/usuario/views/VentanaPago$4.class
	sprint2/target/classes/usuario/views/VentanaPago.class
	sprint2/target/classes/usuario/views/VentanaPrincipal$1.class
	sprint2/target/classes/usuario/views/VentanaPrincipal$2.class
	sprint2/target/classes/usuario/views/VentanaPrincipal$3.class
	sprint2/target/classes/usuario/views/VentanaPrincipal$4.class
	sprint2/target/classes/usuario/views/VentanaPrincipal$RestarBoton.class
	sprint2/target/classes/usuario/views/VentanaPrincipal$SumarBoton.class
	sprint2/target/classes/usuario/views/VentanaPrincipal.class
	sprint2/target/classes/usuario/views/VentanaTarjeta$2.class
	sprint2/target/classes/usuario/views/VentanaTarjeta$3.class
	sprint2/target/classes/usuario/views/VentanaTarjeta.class

+++ sprint2/src/main/java/common/modelo/Pedido.java
@@ -2,7 +2,7 @@ package common.modelo;
 
 public class Pedido {
 	private int idPedido;
-	private int precioTotal;
+	private double precioTotal;
 	private String fecha;
 	private String albaran;
 	private String direccion;
@@ -28,10 +28,10 @@ public class Pedido {
 	public void setIdPedido(int idPedido) {
 		this.idPedido = idPedido;
 	}
-	public int getPrecioTotal() {
+	public double getPrecioTotal() {
 		return precioTotal;
 	}
-	public void setPrecioTotal(int precioTotal) {
+	public void setPrecioTotal(double precioTotal) {
 		this.precioTotal = precioTotal;
 	}
 	public String getFecha() {
